requirements.txt
ipy-table==1.15.1
matplotlib==3.1.0
numpy==1.16.4
pandas==0.24.2
PyYAML==5.1
